Function naming
Functions should be named <subject>_<verb> not the other way arround.
This way, there is no list
create_foo
create_bar
but a list
foo_create
bar_create
which is a lot nicer in my opinion. It is important to use this rule
everywhere. Otherwise the API will be a huge mess.