Build file (Linux) and config.hpp change
To simplify building Aura in parallel, the config.hpp file is now not
generated in the source directory but in the build directory. This is the
correct location for a generated file anyways. An installation script would
need to copy this file to the installation directory of the include files.