More sane printing of OpenCL build log
Formerly, even an empty module build log would be printed (which still
consists of '\n\0'). Now, we check for that sequence of bytes before printing.
This used to cause lots of blank lines printed on the console on projects
which use a lot of kernels.

Furthermore, the build log is always printed in create_module_from_string. So
before, when the debug paramter in load_from_string was true, the build log
was printed a second time. Now it is always printed, and always printed once.