Fix OpenCL memory leak and module rewrite
Formerly, the OpenCL cl_kernel objects were never released. This created a
memory leak for applications which used aura, cleared up all references to it,
and continued running: those applications would continue to use device memory.
This was especially noticable by applications that used aura through Matlabs
mex-interface: Here every invocation of the .mex-file would use up more and
more device memory, until matlab was restarted.

In order to fix this, the cl_kernel objects need to be remembered and cleaned
up using clReleaseKernel(). Since this was not possible in a nice way in the
implementation, a proper module type was created to take care of this.

This new module type abstracts the cl_program type. It takes care of proper
cleanup of both the cl_program and cl_kernel's.

Added feature here is support for multiple kernels in a single module, and
caching of kernels in the module type. (This is stil untested! and only
implemented for OpenCL!)