Update fill_args_ to use proper iterator type
The iterator in fill_args_ was of the wrong type, so the type was relaxed to allow the entire iterator type to be deduced by the compiler. Only caveat is that this will result in less understandable compiler errors if a non-iterator type is passed to fill_args_. The upside is that it works with Visual Studio 2015 :)

Also, standardized the use of std::size_t as template argument types for arg_t and args_tt.